adding new function get_partitions

+++ kernel/src/fs/fat32_fs/include/fat32_mount.h
@@ -5,5 +5,5 @@
 #include <stdbool.h>
 
 
-bool create_fat32_volume( uint64_t start_lba, uint32_t sectors);
+bool create_fat32_volume( uint32_t start_lba, uint32_t sectors);
 bool fat32_mount( uint64_t start_lba);
